fix(events): delete events on clickhouse only if activated (#3324)

+++ app/jobs/billable_metrics/delete_events_job.rb
@@ -27,21 +27,23 @@ module BillableMetrics
       ).update_all(deleted_at:) # rubocop:disable Rails/SkipsModelValidations
 
       # Delete events_raw & events_enriched on clickhouse using `external_subscription_id`
-      Clickhouse::EventsRaw.where(
-        organization_id: metric.organization.id,
-        code: metric.code,
-        external_subscription_id: Charge.with_discarded
-          .where(billable_metric_id: metric.id)
-          .joins(plan: :subscriptions).pluck("subscriptions.external_id")
-      ).delete_all
+      if ENV["LAGO_CLICKHOUSE_ENABLED"].present?
+        Clickhouse::EventsRaw.where(
+          organization_id: metric.organization.id,
+          code: metric.code,
+          external_subscription_id: Charge.with_discarded
+            .where(billable_metric_id: metric.id)
+            .joins(plan: :subscriptions).pluck("subscriptions.external_id")
+        ).delete_all
 
-      Clickhouse::EventsEnriched.where(
-        organization_id: metric.organization.id,
-        code: metric.code,
-        external_subscription_id: Charge.with_discarded
-          .where(billable_metric_id: metric.id)
-          .joins(plan: :subscriptions).pluck("subscriptions.external_id")
-      ).delete_all
+        Clickhouse::EventsEnriched.where(
+          organization_id: metric.organization.id,
+          code: metric.code,
+          external_subscription_id: Charge.with_discarded
+            .where(billable_metric_id: metric.id)
+            .joins(plan: :subscriptions).pluck("subscriptions.external_id")
+        ).delete_all
+      end
     end
   end
 end
